+++ examples/resources/hubspot_crm_user_profile/import.sh
@@ -1,0 +1,1 @@
+tofu import hubspot_crm_user_profile.operator 'membership:123456789'